pyproject.toml
[project]
name = "clip"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "torch==2.2.2",
    "open-clip-torch>=3.2.0",
    "pillow>=12.0.0",
    "pydantic>=2.12.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
    "qdrant-client<1.16.0",
    "pydantic-settings>=2.12.0",
    "transformers>=4.57.3",
    "numpy==1.26.4",
]
